👺 Havoc: Add failing test proving stack overflow in deep AST cloning
Demonstrates a fatal SIGABRT due to stack overflow in `AnalyzedExpr::clone` and `Drop` on deeply nested ASTs.
The tests correctly isolate the SIGABRT in child processes.
